CCPFUCK/requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.9
django-crispy-forms==1.6.1
django-formtools==2.0
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-simple-captcha==0.5.5
djangorestframework==3.7.1
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
jdcal==1.4
MySQL-python==1.2.5
odfpy==1.4.0
olefile==0.44
openpyxl==2.6.0
Pillow==4.2.1
pytz==2017.2
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
